% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aeles}
\alias{aeles}
\title{Australian election study data}
\format{
A data frame with 1,468 rows and 4 variables
\describe{
\item{id}{The id of the respondent}
\item{year}{Year}
\item{party}{Party that the respondent votes for in the House of Representatives}
\item{gender}{Gender of the respondent}
}
}
\source{
This dataset is from the following; Australian Election Study
\url{https://dataverse.ada.edu.au/file.xhtml?fileId=18013&version=3.0}
}
\usage{
aeles
}
\description{
The aes dataset contains the answers to the surveys that were done in 2019 for the election.
This dataset only includes the id, year, party, and gender from the full survey data. The year
column comes from the transformations of two different questions to see whether the voter voted for the
same party in 2016 and 2019 or not, and if not, who did they vote for before?
}
\keyword{datasets}
