% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anim-animate.R
\name{anim_animate}
\alias{anim_animate}
\title{Modified the ggplot object}
\usage{
anim_animate(plot, modify = FALSE)
}
\arguments{
\item{plot}{ggplot object}

\item{modify}{Default is FALSE which supplied the function needed for the rendering
package.}
}
\value{
A gganimate object if the rendering is gganimate or a ggplot object if
the rendering is plotly.
}
\description{
This function by default will modified the ggplot object before the user can
pass it to the rendering of choice.
}
\examples{
data <- prep_anim(data = osiris, id = firmID, values = sales, time = year, ngroup = 5)

label = c("Top 20\%", "21-40", "41-60", "61-80", "81-100", "Not listed", "test")

p <- anim_plot(data = data, id = firmID, time = year, color = japan, label = label)

p2 <- anim_animate(p)

}
