% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anim-plot.R
\name{anim_plot}
\alias{anim_plot}
\title{Turn the data into ggplot object for animate function}
\usage{
anim_plot(
  data,
  palette = RColorBrewer::brewer.pal(9, "Set1"),
  rendering = "gganimate"
)
}
\arguments{
\item{data}{The data frame that has been prepared by the \code{\link[=prep_anim]{prep_anim()}}}

\item{palette}{The vector of palette used by the function to supplied the color for each group.}

\item{rendering}{The choice of method used to create and display the plot, either gganimate or
plotly.}
}
\value{
Return a ggplot object
}
\description{
This function take in the data which has been prepared by the \code{\link[=prep_anim]{prep_anim()}} and
return the ggplot object. The user can still modified the plot the same as normal.
}
\examples{
data <- anim_prep(data = toy_dbl, id = id, values = values, time = year, label = check, color = color)

anim_plot(data)

}
