% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anim-prep.R
\name{anim_prep}
\alias{anim_prep}
\title{Prepare Numerical Data into categorized data}
\usage{
anim_prep(
  data,
  id = NULL,
  values = NULL,
  time = NULL,
  ngroup = 5L,
  breaks = NULL,
  group_scaling = NULL,
  scaling = "rank",
  time_dependent = FALSE,
  color = NULL,
  label = NULL
)
}
\arguments{
\item{data}{A data frame containing the data to be prepared for visualization.}

\item{id}{The column name that represents the unique identifier variable.}

\item{values}{The column name that contains the numeric values to be visualized.}

\item{time}{The column name the represents the time variable.}

\item{ngroup}{The number of groups or categories to create for scaling values.}

\item{breaks}{A vector of breaks for creating bins.}

\item{group_scaling}{The column name that represents the grouping variable.}

\item{scaling}{The scaling method to be used; "rank" or "absolute".}

\item{time_dependent}{Logical. Should the visualization be time-dependent? The default is FALSE}

\item{color}{The column name to be used in \code{\link[ggplot2:aes]{ggplot2::aes()}} for the plot function.}

\item{label}{A vector of labels to represented each values.}
}
\value{
An animbook object:
\item{data}{A data frame in the categorized format}
\item{settings}{A list of settings to be used in the plot function, including gap,
xbreaks, label, scaling, time_dependent}
}
\description{
This function prepares the numerical data into the categorized format by
grouping data, scaling values, and creating necessary data and settings for the plot function.
}
\details{
The function takes the input data and performs several operations to prepare it for visualizations.
It groups data, scales values, and transformed it to categorized data along with the settings
for the plot function.
}
\examples{
anim_prep(data = osiris, id = ID, values = sales, time = year)

anim_prep(data = osiris, id = ID, values = sales, time = year,
group_scaling = country)

anim_prep(data = osiris, id = ID, values = sales, time = year,
scaling = "absolute")

anim_prep(data = osiris, id = ID, values = sales, time = year,
group_scaling = country, scaling = "absolute")

}
