% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cat_change}
\alias{cat_change}
\title{Simulated data with some change (category)}
\format{
A data frame with 400 rows and 4 variables
\describe{
\item{id}{The id of the organisation}
\item{time}{time}
\item{gp}{Either X or Y}
\item{qnt}{Quantile group for the two times}
}
}
\usage{
cat_change
}
\description{
This data has changed from category A to E between two-time points.
}
\examples{
d <- anim_prep_cat(cat_change, id = id, values = qnt,
time = time, group = gp)

d_p <- wallaby_plot(d, height = 1)

d_p_anim <- anim_animate(d_p)
}
\keyword{datasets}
