% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_w_change}
\alias{d_w_change}
\title{Simulated data with some change}
\format{
A data frame with 400 rows and 4 variables
\describe{
\item{id}{The id of the organisation}
\item{time}{time}
\item{gp}{Either X or Y}
\item{qnt}{Quantile group for the two times}
}
}
\usage{
d_w_change
}
\description{
This data has changes from category A to E
between two time points.
}
\examples{
d <- anim_prep_cat(d_w_change, id = id,
values = qnt, time = time,
color = gp, time_dependent = FALSE)

d_p <- wallaby_plot(d, height = 1)

d_p_anim <- anim_animate(d_p)
}
\keyword{datasets}
