% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dbl_change}
\alias{dbl_change}
\title{Simulated data with some change (numerical)}
\format{
A data frame with 400 rows and 4 variables
\describe{
\item{id}{The id of the organisation}
\item{time}{time}
\item{gp}{Either X or Y}
\item{values}{Numerical values represent sales}
}
}
\usage{
dbl_change
}
\description{
This data contained the numerical values for each observation.
}
\examples{
d <- anim_prep(dbl_change, id = id, values = values,
time = time, color = gp, time_dependent = FALSE)

d_p <- wallaby_plot(d, height = 1)

d_p_anim <- anim_animate(d_p)
}
\keyword{datasets}
