% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet-plot.R
\name{facet_plot}
\alias{facet_plot}
\title{Line Plot with Facets}
\usage{
facet_plot(data, palette = RColorBrewer::brewer.pal(9, "Set1"))
}
\arguments{
\item{data}{A data frame containing the input data.}

\item{palette}{The vector of palette used by the function to supplied the color for each group.}
}
\value{
Return a ggplot object
}
\description{
This function creates a line plot with facets based on time periods and grouping variable.
}
\examples{
data <- anim_prep(data = osiris, id = firmID, values = sales, time = year)

facet_plot(data)

}
