% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kangaroo-data.R
\name{kangaroo_draw}
\alias{kangaroo_draw}
\title{Kangaroo path map}
\usage{
kangaroo_draw(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A mapped data
}
\description{
This function is used to interpolate the path for the kangaroo plot.
}
\keyword{internal}
