% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{osiris}
\alias{osiris}
\title{Osiris firm sales data}
\format{
A data frame with 10,270 rows and 5 variables
\describe{
\item{year}{Year}
\item{ID}{BvD(Bureau van Dijk) ID}
\item{country}{Address of incorp. - Country}
\item{sales}{Sales}
\item{japan}{Whether the firm is froom Japan or not}
}
}
\source{
This dataset is from the following; Bureau van Dijk
\url{https://www.bvdinfo.com/en-gb/our-products/data/international/osiris}.
}
\usage{
osiris
}
\description{
The Osiris dataset contains information on listed, and major unlisted/delisted
companies across the world from 2006 to 2018. This dataset only includes the year, ID,
country, sales, and japan variables from the full Osiris dataset to give the user
an example of the dataset format.
}
\keyword{datasets}
