% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep-anim.R
\name{prep_anim}
\alias{prep_anim}
\title{Prep the data structure for plot function}
\usage{
prep_anim(data, id = NULL, values = NULL, time = NULL, ngroup = 5)
}
\arguments{
\item{data}{The data frame or a tibble which contains the information of interested.}

\item{id}{Name of the column that uniquely identified the elements of interested}

\item{values}{Name of the column of interested, must be a numeric column.}

\item{time}{Name of the columns that will be used as time variable.}

\item{ngroup}{Number of group the user want to split the values into (excluded
the NA group which is the additional group).}
}
\value{
Return a tibble with additional frame and qtile columns (excluded the
values columns which been used to calculated the qtile column)
}
\description{
This function prepared the data into the format the \code{\link[=anim_plot]{anim_plot()}} required.
}
\examples{
prep_anim(data = osiris, id = firmID, values = sales, time = year, ngroup = 5)

}
