% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey-shade.R
\name{sankey_shade}
\alias{sankey_shade}
\title{Sankey shade data}
\usage{
sankey_shade(initial = NULL, proportion = NULL, y = NULL, position = "left")
}
\arguments{
\item{initial}{A numerical of initial values for the Sankey shade.}

\item{proportion}{A vector of proportions corresponding to each group.}

\item{y}{A numeric vector specifying the vertical position of the sankey diagram.}

\item{position}{A character string specifying the side the of the Sankey diagram.
This need to be either "left" (default) or "right"}
}
\value{
A data frame of the shaded Sankey diagram.
}
\description{
This function creates a shaded Sankey using the specified parameters.
}
\examples{
data <- sankey_shade(4, c(0.4, 0.3, 0.2, 0.1), c(4, 3, 2, 1))

ggplot2::ggplot(data = data, ggplot2::aes(x = x, y = y, group = id, fill = as.factor(id))) +
  ggplot2::geom_polygon(alpha = 0.1)

}
\keyword{internal}
