% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anim-animate.R
\name{anim_animate}
\alias{anim_animate}
\title{Modified the ggplot object}
\usage{
anim_animate(plot)
}
\arguments{
\item{plot}{ggplot object}
}
\value{
A gganimate object if the rendering is gganimate or a ggplot object if
the rendering is plotly.
}
\description{
This function will modify the ggplot object before the user can
pass it to the rendering of choice.
}
\examples{
animbook <- anim_prep(data = osiris, id = ID, values = sales, time = year, group = japan)

plot <- wallaby_plot(animbook)

animate <- anim_animate(plot)

plotly::ggplotly(animate)

}
