% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anim-prep.R
\name{anim_prep}
\alias{anim_prep}
\title{Prepare Data for Animated Visualizations}
\usage{
anim_prep(
  data,
  id = NULL,
  values = NULL,
  time = NULL,
  label = NULL,
  ngroup = 5L,
  breaks = NULL,
  group_scaling = NULL,
  color = NULL,
  time_dependent = TRUE,
  scaling = "rank",
  runif_min = 1,
  runif_max = 50
)
}
\arguments{
\item{data}{A data frame containing the data to be prepared for visualization.}

\item{id}{The column name that represents the unique identifier variable.}

\item{values}{The column name that contains the numeric values to be visualized.}

\item{time}{The column name representing the time variable.}

\item{label}{A vector of labels to be used for the y-axis in the visualization.}

\item{ngroup}{The number of groups or categories to create for scaling values.}

\item{breaks}{A vector of breaks for creating bins when using absolute scaling.}

\item{group_scaling}{The column name that represents the grouping variable.}

\item{color}{The column name to used in \code{\link[=aes]{aes()}} for the \code{\link[=anim_plot]{anim_plot()}}.}

\item{time_dependent}{Logical. Should the visualization be time-dependent? Default is TRUE.}

\item{scaling}{The scaling method to be used: "rank" or "absolute".}

\item{runif_min}{The minimum value for random addition to frame numbers.}

\item{runif_max}{The maximum value for random addition to frame numbers.}
}
\value{
A list containing the following components:
\item{data}{A data frame with prepared data for visualization.}
\item{rect_data}{A data frame that wil be used for shading in \code{\link[=anim_plot]{anim_plot()}}.}
\item{settings}{A list of settings to be used in \code{\link[=anim_plot]{anim_plot()}}, including gap, breaks, labels and scaling.}
}
\description{
This function prepares the data into the format the \code{\link[=anim_plot]{anim_plot()}} required by assigning frames,
grouping data, scaling values, and creating a necessary data and settings for the \code{\link[=anim_plot]{anim_plot()}} function.
}
\examples{
anim_prep(data = osiris, id = firmID, values = sales, time = year)

anim_prep(data = osiris, id = firmID, values = sales, time = year, group_scaling = country)

anim_prep(data = osiris, id = firmID, values = sales, time = year, scaling = "absolute")

anim_prep(data = osiris, id = firmID, values = sales, time = year, group_scaling = country, scaling = "absolute")

}
