% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anim-prep-cat.R
\name{anim_prep_cat}
\alias{anim_prep_cat}
\title{Prepare Category Data for Visualizations}
\usage{
anim_prep_cat(
  data,
  id = NULL,
  values = NULL,
  time = NULL,
  label = NULL,
  order = NULL,
  color = NULL,
  time_dependent = TRUE,
  runif_min = 1,
  runif_max = 50
)
}
\arguments{
\item{data}{A data frame containing the category values to be prepared for visualization.}

\item{id}{The column name that represents the unique identifier variable.}

\item{values}{The column name that contains the categorical values to be visualized.}

\item{time}{The column name represents the time variable.}

\item{label}{A vector of labels to be used for the y-axis in the visualization.}

\item{order}{A vector of order for sorting the category values.}

\item{color}{The column name to be used in \code{\link[ggplot2:aes]{ggplot2::aes()}} for the plot function.}

\item{time_dependent}{Logical. Should the visualization be time-dependent? Default is TRUE.}

\item{runif_min}{The minimum value for random addition to frame numbers.}

\item{runif_max}{The maximum value for random addition to frame numbers.}
}
\value{
An animbook object:
\item{data}{A data frame with prepared data for visualization.}
\item{settings}{A list of settings to be used in plot function, including gap, xbreaks, label, scaling, time_dependent,
runif_min, and runif_max}
}
\description{
This function prepares the category data into the format the plot function required
by assigning frames and creating necessary data and settings for the plot function.
}
\details{
The function takes the input data and performs several operations to prepare it for visualizations.
It assigns frames and creates necessary data and settings for the plot function.
}
\examples{
anim_prep_cat(data = aeles, id = id, values = party, time = year)

}
