% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funnel-web-data.R
\name{funnel_web_spider_data}
\alias{funnel_web_spider_data}
\title{Funnel web spider plot data}
\usage{
funnel_web_spider_data(data)
}
\arguments{
\item{object}{An animbook object.}
}
\value{
A modified animbook object.
}
\description{
This function performs data manipulation for facetting.
}
\details{
The function takes the animbook object and manipulates the data into the
format where it can be further facetting.
}
\keyword{internal}
