% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kangaroo-plot.R
\name{kangaroo_data}
\alias{kangaroo_data}
\title{Kangaroo plot data}
\usage{
kangaroo_data(object, height = 0.6, width = 50)
}
\arguments{
\item{object}{An animbook object}
}
\value{
A modified animbook object with additional data components
}
\description{
This function performs data manipulation and formatting tasks
from the original object with additional data components for labeling and shading.
}
\details{
The function takes the animbook object and creates a new label data and
shading data used for the plot then appends them to the original object.
}
\keyword{internal}
