% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kangaroo-data.R
\name{kangaroo_data}
\alias{kangaroo_data}
\title{Kangaroo plot data}
\usage{
kangaroo_data(data, time_dependent = FALSE, height = 0.6, width = 50L)
}
\arguments{
\item{data}{The categorized data.}

\item{time_dependent}{Logical. Should the visualization be time-dependent?
The default is FALSE.}

\item{height}{The proportion the point takes in the shaded area.}

\item{width}{The number that controls the runif_max to specify how far apart
each point is.}
}
\value{
An object contained the modified data with additional data components.
}
\description{
This function performs data manipulation and path interpolation of the
categorized data. This includes additional data components for labeling
and shading.
}
\details{
The function takes the categorized object and interpolates the path
for each observation. Additionally, the label and shading data are created.
}
\keyword{internal}
