% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kangaroo-plot.R
\name{kangaroo_plot}
\alias{kangaroo_plot}
\title{Turn the data into a ggplot object for the animate function}
\usage{
kangaroo_plot(
  object,
  palette = RColorBrewer::brewer.pal(9, "Set1"),
  rendering = "ggplot",
  ...
)
}
\arguments{
\item{object}{The animbook object returned from the prep function.}

\item{palette}{The vector of the palette used by the function to supply the color to each group.}

\item{rendering}{The choice of method used to create and display the plot, either gganimate or
plotly.}

\item{...}{Additional arguments for customization, see details for more information}
}
\value{
Return a ggplot object
}
\description{
This function takes in the data which has been prepared by the \code{\link[=anim_prep]{anim_prep()}} or \code{\link[=anim_prep_cat]{anim_prep_cat()}}
and return the ggplot object. The user can still modify the plot the same as normal.
}
\details{
This function takes prepared data and generates a ggplot object.
The kangaroo plot is the plot that shows the movement between groups over time.
The jitter point can be controlled using additional arguments such as height and width
to control the appearance. For the shading area, the alpha argument can be used.
}
\examples{
animbook <- anim_prep(data = osiris, id = ID, values = sales, time = year, color = japan)

kangaroo_plot(animbook)

}
