% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osiris.R
\docType{data}
\name{osiris}
\alias{osiris}
\title{Osiris firm sales data}
\format{
A data frame with 7,270 rows and 4 variables
\describe{
\item{year}{Year}
\item{firmID}{Firm name + BvD(Bureau van Dijk) ID}
\item{country}{Address of incorp. - Country}
\item{sales}{Sales}
}
}
\source{
This dataset is from the following; Bureau van Dijk
\url{https://www.bvdinfo.com/en-gb/our-products/data/international/osiris}.
}
\usage{
osiris
}
\description{
The Osiris dataset contains information on listed, and major unlisted/delisted
companies across the world from 2009-2018. This dataset only includes the year,
firmID, country, and sales variables from the full Osiris dataset to give the user
an example of the dataset format.
}
\keyword{datasets}
