% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallaby-plot.R
\name{wallaby_data}
\alias{wallaby_data}
\title{Wallaby plot data}
\usage{
wallaby_data(object, subset = "top", relation = "one_many")
}
\arguments{
\item{object}{An animbook object}
}
\value{
A modified animbook object with addition data components
}
\description{
This function performs data manipulation and formatting tasks
from the original object with additional data components for labeling and shading.
}
\details{
The function takes the animbook object then subset the data based on the users.
Additionaly, it create a label data and shading data for the \code{\link[=anim_plot]{anim_plot()}} function. All
of this then replaced the original data and append new data to the object.
}
\keyword{internal}
